+++ core/dbt/parser/read_files.py
@@ -5,7 +5,7 @@ from dbt.contracts.files import (
 )
 
 from dbt.parser.schemas import yaml_from_file, schema_file_keys, check_format_version
-from dbt.exceptions import CompilationException
+from dbt.exceptions import ParsingException
 from dbt.parser.search import filesystem_search
 from typing import Optional
 
@@ -54,17 +54,17 @@ def validate_yaml(file_path, dct):
             if not isinstance(dct[key], list):
                 msg = (f"The schema file at {file_path} is "
                        f"invalid because the value of '{key}' is not a list")
-                raise CompilationException(msg)
+                raise ParsingException(msg)
             for element in dct[key]:
                 if not isinstance(element, dict):
                     msg = (f"The schema file at {file_path} is "
                            f"invalid because a list element for '{key}' is not a dictionary")
-                    raise CompilationException(msg)
+                    raise ParsingException(msg)
                 if 'name' not in element:
                     msg = (f"The schema file at {file_path} is "
                            f"invalid because a list element for '{key}' does not have a "
                            "name attribute.")
-                    raise CompilationException(msg)
+                    raise ParsingException(msg)
 
 
 # Special processing for big seed files