+++ core/dbt/adapters/base/relation.py
@@ -89,7 +89,10 @@ class BaseRelation(FakeAPIObject, Hashable):
             if not self._is_exactish_match(k, v):
                 exact_match = False
 
-            if self.path.get_lowered_part(k) != v.lower():
+            if (
+                self.path.get_lowered_part(k).strip(self.quote_character) !=
+                v.lower().strip(self.quote_character)
+            ):
                 approximate_match = False
 
         if approximate_match and not exact_match: